+++ CHANGELOG.md
@@ -14,6 +14,7 @@ x.x.x Release notes (yyyy-MM-dd)
 
 ### Internal
 * Upgraded the React Native integration tests app (now using RN v0.61.3). ([#2603](https://github.com/realm/realm-js/pull/2603))
+* Added a test to verify that an exception is thrown when an object schema has no properties.
 
 3.4.2 Release notes (2019-11-14)
 =============================================================